backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1.1
django-allauth==64.2.0
django-cors-headers==4.4.0
django-filter==23.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
Markdown==3.5.1
msgpack==1.1.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
psycopg2==2.9.9
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
ShopifyAPI==12.6.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.3
